% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxVarElement.R
\docType{methods}
\name{-,ANY,AuxVarElementClass-method}
\alias{-,ANY,AuxVarElementClass-method}
\title{Substrac element.}
\usage{
\S4method{-}{ANY,AuxVarElementClass}(e1, e2)
}
\arguments{
\item{e1}{AuxVarElementClass.}

\item{e2}{ANY.}
}
\description{
Substrac element.
}
