% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substract.R
\docType{methods}
\name{-,ANY,VarExpressionClass-method}
\alias{-,ANY,VarExpressionClass-method}
\title{Substract ANY object with VarExpression object.}
\usage{
\S4method{-}{ANY,VarExpressionClass}(e1, e2)
}
\arguments{
\item{e1}{ANY.}

\item{e2}{VarExpressionClass.}
}
\description{
Substract ANY object with VarExpression object.
}
