% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxVar.R
\name{AuxVar}
\alias{AuxVar}
\title{Function to create auxiliar variables.}
\usage{
AuxVar(name, expr, iterator = list(), description = "")
}
\arguments{
\item{name}{variable name.}

\item{expr}{variable expression.}

\item{iterator}{iterator to build the collection of variables.}

\item{description}{variable description.}

\item{start_position}{starting point to enumerate the set of variables.}
}
\description{
Function to create auxiliar variables.
}
