% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxVar.R
\docType{class}
\name{AuxVarClass-class}
\alias{AuxVarClass-class}
\alias{AuxVarClass}
\title{Auxiliary Variable class.}
\description{
Auxiliary Variable class.
}
\section{Slots}{

\describe{
\item{\code{name}}{character.}

\item{\code{sets}}{list.}

\item{\code{position}}{arrayORnumeric.}

\item{\code{variable}}{list.}

\item{\code{description}}{character.}
}}

