% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxVarElement.R
\docType{methods}
\name{-,AuxVarElementClass,ANY-method}
\alias{-,AuxVarElementClass,ANY-method}
\alias{-,AuxVarElementClass,ANY-method}
\title{Substrac element.}
\usage{
\S4method{-}{AuxVarElementClass,ANY}(e1, e2)

\S4method{-}{AuxVarElementClass,ANY}(e1, e2)
}
\arguments{
\item{e1}{AuxVarElementClass.}

\item{e2}{ANY.}

\item{e1}{AuxVarElementClass.}
}
\description{
Substrac element.

Substrac element.
}
