% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_var_element.R
\docType{class}
\name{AuxVarElementClass-class}
\alias{AuxVarElementClass-class}
\alias{AuxVarElementClass}
\title{Auxiliary variable element class.}
\description{
Auxiliary variable element class.
}
\section{Slots}{

\describe{
\item{\code{name}}{character.}

\item{\code{expr}}{VarExpressionClass.}

\item{\code{description}}{character.}
}}

