% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{$<-,ModelClass-method}
\alias{$<-,ModelClass-method}
\title{Assign objects to model class.}
\usage{
\S4method{$}{ModelClass}(x, name) <- value
}
\arguments{
\item{ModelClass}{}
}
\description{
Assign objects to model class.
}
