% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraint.R
\name{Constraint}
\alias{Constraint}
\title{Function to build constraints.}
\usage{
Constraint(name, expr, sets = list(), start_position = 1,
  description = "")
}
\arguments{
\item{name}{constraint name.}

\item{expr}{mathematical expression to build the constraints.}

\item{sets}{set to build the collection of constraints.}

\item{start_position}{starting point to enumerate the set of constraints.}

\item{description}{constraints description.}
}
\value{
Object of ConstraintClass or ConstraintElementClass class .
}
\description{
Function to build constraints.
}
\examples{
D <- Var("D")
Constraint("Demand", D <= 2, description = "max demand.")
}
