% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraint.R
\docType{class}
\name{ConstraintClass-class}
\alias{ConstraintClass-class}
\alias{ConstraintClass}
\title{Constraint class.}
\value{
Object of ConstraintClass class.
}
\description{
Constraint class.
}
\section{Slots}{

\describe{
\item{\code{name}}{character.}

\item{\code{sets}}{list.}

\item{\code{position}}{arrayORnumeric.}

\item{\code{constraint}}{list.}

\item{\code{description}}{character.}
}}

