% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstraintElement.R
\docType{class}
\name{ConstraintElementClass-class}
\alias{ConstraintElementClass-class}
\alias{ConstraintElementClass}
\title{Constraint element class.}
\value{
object of the ConstraintElementClass class.
}
\description{
Constraint element class.
}
\section{Slots}{

\describe{
\item{\code{name}}{character.}

\item{\code{position}}{numeric.}

\item{\code{expr}}{numeric.}

\item{\code{state}}{character.}

\item{\code{description}}{character.}
}}

