% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstraintExpression.R
\name{ConstraintExpression}
\alias{ConstraintExpression}
\title{Function to build a constraint expression.}
\usage{
ConstraintExpression(lhs, sense, rhs)
}
\arguments{
\item{lhs}{left hand side.}

\item{sense}{sense. Possibilities c("<=", ">=", "==").}

\item{rhs}{right hand side.}
}
\description{
Function to build a constraint expression.
}
