% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{get_indexed_expr}
\alias{get_indexed_expr}
\title{Substitute index by its value.}
\usage{
get_indexed_expr(string, iterator, i, ind)
}
\arguments{
\item{string}{string with mathematical expression.}

\item{iterator}{iterator.}

\item{i}{combination of the indices name.}

\item{ind}{object with combination indices values.}
}
\description{
Substitute index by its value.
}
