% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{\%inset\%}
\alias{\%inset\%}
\title{Operator that return the considered name of the element and the iterator.}
\usage{
i \%inset\% s
}
\arguments{
\item{i}{element name.}

\item{s}{SetClass.}
}
\description{
Operator that return the considered name of the element and the iterator.
}
