% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstraintExpression.R
\docType{methods}
\name{>=,VarExpressionClass,VarElementClass-method}
\alias{>=,VarExpressionClass,VarElementClass-method}
\title{Great or equal method to compare variable expressions and variable element.}
\usage{
\S4method{>=}{VarExpressionClass,VarElementClass}(e1, e2)
}
\arguments{
\item{e1}{VarExpressionClass.}

\item{e2}{VarElementClass.}
}
\description{
Great or equal method to compare variable expressions and variable element.
}
