% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstraintExpression.R
\docType{methods}
\name{>=,VarExpressionClass,VarExpressionClass-method}
\alias{>=,VarExpressionClass,VarExpressionClass-method}
\title{Great or equal method to compare two variable expressions.}
\usage{
\S4method{>=}{VarExpressionClass,VarExpressionClass}(e1, e2)
}
\arguments{
\item{e1}{VarExpressionClass.}

\item{e2}{VarExpressionClass.}
}
\description{
Great or equal method to compare two variable expressions.
}
