% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListSets.R
\docType{methods}
\name{indices,list-method}
\alias{indices,list-method}
\title{Function to obtain the combination of the elements of a set of sets.}
\usage{
\S4method{indices}{list}(object)
}
\arguments{
\item{list}{}
}
\description{
Function to obtain the combination of the elements of a set of sets.
}
