% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sets.R
\name{indices}
\alias{indices}
\title{Function to obtain the combination of the elements of a set of sets.}
\usage{
indices(object)
}
\arguments{
\item{object}{}
}
\description{
Function to obtain the combination of the elements of a set of sets.
}
