% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{Iter}
\alias{Iter}
\title{Function to build a list of iterators object.}
\usage{
Iter(...)
}
\arguments{
\item{...}{objects of Iterator class.}
}
\value{
list with iterator objects.
}
\description{
Function to build a list of iterators object.
}
