% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\docType{class}
\name{IteratorClass-class}
\alias{IteratorClass-class}
\alias{IteratorClass}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{i}}{ANY.}

\item{\code{set}}{SetClass.}
}}

