% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sets.R
\name{ListSets}
\alias{ListSets}
\title{Function to obtain a set of sets.}
\usage{
ListSets(...)
}
\arguments{
\item{...}{undefined number of set object from the class SetClass.}
}
\description{
Function to obtain a set of sets.
}
