% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Model}
\alias{Model}
\title{Initialize model object.}
\usage{
Model()
}
\description{
Initialize model object.
}
\note{
Work with envioronments to simplify notation of constraints and to avoid the use of eval.parent
}
