% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{class}
\name{ModelClass-class}
\alias{ModelClass-class}
\alias{ModelClass}
\title{Model class.}
\description{
Model class.
}
\section{Slots}{

\describe{
\item{\code{objects}}{list.}

\item{\code{info}}{ModelInfoClass.}
}}

