% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfo.R
\docType{class}
\name{ModelInfoClass-class}
\alias{ModelInfoClass-class}
\alias{ModelInfoClass}
\title{Model information class.}
\description{
Model information class.
}
\section{Slots}{

\describe{
\item{\code{nvars}}{numeric.}

\item{\code{nobjs}}{numeric.}

\item{\code{ncons}}{numeric.}
}}

