% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\docType{class}
\name{ObjectiveClass-class}
\alias{ObjectiveClass-class}
\alias{ObjectiveClass}
\title{Objective function class.}
\description{
Objective function class.
}
\section{Slots}{

\describe{
\item{\code{name}}{objective functin name.}

\item{\code{expr}}{mathematical expression.}

\item{\code{sense}}{optimization sense. Possibilities c("maximize", "minimize").}

\item{\code{state}}{objective state.  Possibilities c("active", "deactive").}

\item{\code{description}}{objective function description.}
}}

