% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{Param}
\alias{Param}
\title{Function to create a parameter.}
\usage{
Param(name, value, sets = list(), description = "")
}
\arguments{
\item{name}{parameter name.}

\item{value}{parameter value.}

\item{sets}{set to build the collection of parameters.}

\item{description}{parameter description.}
}
\value{
Object of ParamClass class.
}
\description{
Function to create a parameter.
}
