% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\docType{class}
\name{ParamClass-class}
\alias{ParamClass-class}
\alias{ParamClass}
\title{Parameter class}
\description{
Parameter class
}
\section{Slots}{

\describe{
\item{\code{name}}{character.}

\item{\code{sets}}{list.}

\item{\code{value}}{arrayORnumeric.}

\item{\code{description}}{character.}
}}

