% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\docType{methods}
\name{+,ANY,VarElementClass-method}
\alias{+,ANY,VarElementClass-method}
\title{Sum VarElement object with VarElement object.}
\usage{
\S4method{+}{ANY,VarElementClass}(e1, e2)
}
\arguments{
\item{e1}{ANY.}

\item{e2}{VarElementClass.}
}
\description{
Sum VarElement object with VarElement object.
}
