% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_var_element.R
\docType{methods}
\name{+,AuxVarElementClass,AuxVarElementClass-method}
\alias{+,AuxVarElementClass,AuxVarElementClass-method}
\title{Sum elements.}
\usage{
\S4method{+}{AuxVarElementClass,AuxVarElementClass}(e1, e2)
}
\arguments{
\item{e1}{AuxVarElementClass.}

\item{e2}{AuxVarElementClass.}
}
\description{
Sum elements.
}
