% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\docType{methods}
\name{+,numeric,VarElementClass-method}
\alias{+,numeric,VarElementClass-method}
\title{Sum numeric object with VarElement object.}
\usage{
\S4method{+}{numeric,VarElementClass}(e1, e2)
}
\arguments{
\item{e1}{numeric.}

\item{e2}{VarElementClass.}
}
\description{
Sum numeric object with VarElement object.
}
