% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\docType{methods}
\name{+,numeric,VarExpressionClass-method}
\alias{+,numeric,VarExpressionClass-method}
\title{Sum numeric object with VarExpression object.}
\usage{
\S4method{+}{numeric,VarExpressionClass}(e1, e2)
}
\arguments{
\item{e1}{numeric.}

\item{e2}{VarExpressionClass.}
}
\description{
Sum numeric object with VarExpression object.
}
