% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Addition.R
\docType{methods}
\name{+,VarElementClass,numeric-method}
\alias{+,VarElementClass,numeric-method}
\title{Sum VarElement object with numeric object.}
\usage{
\S4method{+}{VarElementClass,numeric}(e1, e2)
}
\arguments{
\item{e1}{VarElementClass.}

\item{e2}{numeric.}
}
\description{
Sum VarElement object with numeric object.
}
