% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\docType{methods}
\name{+,VarExpressionClass,numeric-method}
\alias{+,VarExpressionClass,numeric-method}
\title{Sum VarExpression object with numeric object.}
\usage{
\S4method{+}{VarExpressionClass,numeric}(e1, e2)
}
\arguments{
\item{e1}{VarExpressionClass.}

\item{e2}{numeric.}
}
\description{
Sum VarExpression object with numeric object.
}
