% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Addition.R
\docType{methods}
\name{+,VarExpressionClass,VarElementClass-method}
\alias{+,VarExpressionClass,VarElementClass-method}
\title{Sum VarExpression object with VarElement object.}
\usage{
\S4method{+}{VarExpressionClass,VarElementClass}(e1, e2)
}
\arguments{
\item{e1}{VarExpressionClass.}

\item{e2}{VarElementClass.}
}
\description{
Sum VarExpression object with VarElement object.
}
