% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{Set}
\alias{Set}
\title{Function to build a set.}
\usage{
Set(name, elements, description = "")
}
\arguments{
\item{name}{set name.}

\item{elements}{set elements.}

\item{description}{set description.}
}
\description{
Function to build a set.
}
