% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\docType{class}
\name{SetClass-class}
\alias{SetClass-class}
\alias{SetClass}
\title{Set class.}
\description{
Set class.
}
\section{Slots}{

\describe{
\item{\code{name}}{character.}

\item{\code{elements}}{vector.}

\item{\code{description}}{character.}
}}

