% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{Solve}
\alias{Solve}
\title{Solve model.}
\usage{
Solve(model, solver = "gurobi", solver_options = list())
}
\arguments{
\item{model}{model of ModelClass class.}

\item{solver}{solver name. Default to 'gurobi'.}

\item{solver_options}{list with solver options. By default no options.}
}
\description{
Solve model.
}
