% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication.R
\docType{methods}
\name{*,VarExpressionClass,VarElementClass-method}
\alias{*,VarExpressionClass,VarElementClass-method}
\title{Title}
\usage{
\S4method{*}{VarExpressionClass,VarElementClass}(e1, e2)
}
\arguments{
\item{e1}{VarExpressionClass.}

\item{e2}{VarElementClass.}
}
\description{
Title
}
