% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Var.R
\name{Var}
\alias{Var}
\title{Function to create variables.}
\usage{
Var(name, sets = list(), start_position = 1, value = NaN, lb = -Inf,
  ub = Inf, state = "unfix", type = "continuous", description = "")
}
\arguments{
\item{name}{variable name.}

\item{sets}{set to build the collection of variables.}

\item{start_position}{starting point to enumerate the set of variables.}

\item{value}{variable value.}

\item{lb}{variable lower bound.}

\item{ub}{variable upper bound.}

\item{state}{variable state.}

\item{type}{variable type.}

\item{description}{variable description.}
}
\description{
Function to create variables.
}
