% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.R
\docType{class}
\name{VarClass-class}
\alias{VarClass-class}
\alias{VarClass}
\title{Variable class.}
\description{
Variable class.
}
\section{Slots}{

\describe{
\item{\code{name}}{character.}

\item{\code{sets}}{list.}

\item{\code{position}}{arrayORnumeric.}

\item{\code{variable}}{list.}

\item{\code{description}}{character.}
}}

