% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substract.R
\docType{methods}
\name{-,VarElementClass,ANY-method}
\alias{-,VarElementClass,ANY-method}
\title{Substract VarElement object with ANY object.}
\usage{
\S4method{-}{VarElementClass,ANY}(e1, e2)
}
\arguments{
\item{e1}{VarElementClass.}

\item{e2}{ANY.}
}
\description{
Substract VarElement object with ANY object.
}
