% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarElement.R
\docType{class}
\name{VarElementClass-class}
\alias{VarElementClass-class}
\alias{VarElementClass}
\title{Var element class.}
\description{
Var element class.
}
\section{Slots}{

\describe{
\item{\code{name}}{character.}

\item{\code{position}}{numeric.}

\item{\code{value}}{numeric.}

\item{\code{lb}}{numeric.}

\item{\code{ub}}{numeric.}

\item{\code{state}}{character.}

\item{\code{type}}{character.}

\item{\code{description}}{character.}
}}

