% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarExpression.R
\name{VarExpression}
\alias{VarExpression}
\title{Variable expression.}
\usage{
VarExpression(independent = 0, variables = c(NA))
}
\arguments{
\item{independent}{independent term of the variable expression.}

\item{variables}{variables coefficients of the variable expression.}
}
\description{
Variable expression.
}
