% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substract.R
\docType{methods}
\name{-,VarExpressionClass,ANY-method}
\alias{-,VarExpressionClass,ANY-method}
\title{Substract VarExpression object with ANY object.}
\usage{
\S4method{-}{VarExpressionClass,ANY}(e1, e2)
}
\arguments{
\item{e1}{VarExpressionClass.}

\item{e2}{ANY.}
}
\description{
Substract VarExpression object with ANY object.
}
