% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_expression.R
\docType{class}
\name{VarExpressionClass-class}
\alias{VarExpressionClass-class}
\alias{VarExpressionClass}
\title{Variable expression class.}
\description{
Variable expression class.
}
\section{Slots}{

\describe{
\item{\code{independent}}{numeric.}

\item{\code{variables}}{numeric.}
}}

