% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Substract.R
\docType{methods}
\name{-,VarExpressionClass,missing-method}
\alias{-,VarExpressionClass,missing-method}
\title{Substract VarExpression object with VarExpression object.}
\usage{
\S4method{-}{VarExpressionClass,missing}(e1, e2)
}
\arguments{
\item{e1}{missing.}

\item{e2}{VarExpressionClass.}
}
\description{
Substract VarExpression object with VarExpression object.
}
