% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substract.R
\docType{methods}
\name{-,VarExpressionClass,VarExpressionClass-method}
\alias{-,VarExpressionClass,VarExpressionClass-method}
\title{Substract VarExpression object with VarExpression object.}
\usage{
\S4method{-}{VarExpressionClass,VarExpressionClass}(e1, e2)
}
\arguments{
\item{e1}{VarExpressionClass.}

\item{e2}{VarExpressionClass.}
}
\description{
Substract VarExpression object with VarExpression object.
}
