% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_expression.R
\docType{class}
\name{ConstraintExpressionClass-class}
\alias{ConstraintExpressionClass-class}
\alias{ConstraintExpressionClass}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{lhs}}{numeric.}

\item{\code{sense}}{character.}

\item{\code{rhs}}{numeric.}
}}

