% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\docType{class}
\name{ExpressionClass-class}
\alias{ExpressionClass-class}
\alias{ExpressionClass}
\title{Expression class.}
\description{
Expression class.
}
\section{Slots}{

\describe{
\item{\code{expression}}{character.}
}}

