% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\docType{methods}
\name{+,VarElementClass,VarExpressionClass-method}
\alias{+,VarElementClass,VarExpressionClass-method}
\title{Sum VarElement object with VarExpression object.}
\usage{
\S4method{+}{VarElementClass,VarExpressionClass}(e1, e2)
}
\arguments{
\item{e1}{VarElementClass.}

\item{e2}{VarExpressionClass.}
}
\description{
Sum VarElement object with VarExpression object.
}
